// dist/commonjs/renderer/texture/Texture.d.ts
export declare class Texture {
    index: number;
}
export declare var createTexture: () => any;
export declare var disposeTexture: Function;
export declare var getTextureSource: Function;
export declare var setTextureSource: Function;
export declare var getTextureWidth: Function;
export declare var setTextureWidth: Function;
export declare var getTextureHeight: Function;
export declare var setTextureHeight: Function;
export declare var getTextureIsNeedUpdate: Function;
export declare var setTextureIsNeedUpdate: Function;